requirements.txt
crewai
openai
jira
PyGithub
playwright
python-dotenv
